Hackathon_0/Gold/requirements.txt
# Python dependencies for AI Employee - Gold Tier

# Core dependencies
python-dotenv>=1.0.0
pathlib>=1.0.1

# Logging and monitoring
colorlog>=6.8.0

# Gmail API
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# WhatsApp automation
playwright>=1.40.0

# Social media APIs
facebook-sdk>=3.1.0
python-twitter>=3.5
tweepy>=4.14.0
python-linkedin-v2>=0.1.0

# Odoo integration
xmlrpc>=1.0.1

# Web automation
selenium>=4.16.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data processing
pandas>=2.1.4
python-dateutil>=2.8.2

# File handling
watchdog>=3.0.0
markdown>=3.5.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Utilities
pytz>=2023.3
schedule>=1.2.0
